Move header components to Header directory
