+++ apps/smart-forms-react-app/src/components/Dashboard/DashboardHeader/DashboardHeader.tsx
@@ -4,11 +4,11 @@ import { StyledRoot, StyledToolbar } from '../../StyledComponents/Header.styles'
 import React from 'react';
 import FaceIcon from '@mui/icons-material/Face';
 import MedicalServicesIcon from '@mui/icons-material/MedicalServices';
-import AccountPopover from '../../Account/AccountPopover';
+import AccountPopover from '../../Header/AccountPopover';
 import { useTheme } from '@mui/material/styles';
-import PatientPopoverMenu from '../../Account/PatientPopoverMenu';
-import UserPopoverMenu from '../../Account/UserPopoverMenu';
-import UserHeader from '../../Account/UserHeader';
+import PatientPopoverMenu from '../../Header/PatientPopoverMenu';
+import UserPopoverMenu from '../../Header/UserPopoverMenu';
+import UserHeader from '../../Header/UserHeader';
 import useResponsive from '../../../custom-hooks/useResponsive';
 import Logo from '../../Misc/Logo';
